# common/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: TCC PostgreSQL Backend Benchmark API
  version: 0.1.0
  description: API contract shared by all benchmark implementations.
servers:
  - url: http://127.0.0.1:8000
paths:
  /health:
    get:
      operationId: getHealth
      responses:
        "200":
          description: API is available.
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    example: ok
  /customers:
    get:
      operationId: listCustomers
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: Paginated customers.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerPage"
        "400":
          $ref: "#/components/responses/ValidationError"
    post:
      operationId: createCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerRequest"
      responses:
        "201":
          description: Customer created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/ConflictError"
  /customers/{id}:
    get:
      operationId: getCustomer
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Customer found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFoundError"
    put:
      operationId: updateCustomer
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerRequest"
      responses:
        "200":
          description: Customer updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /products:
    get:
      operationId: listProductsByCategory
      parameters:
        - in: query
          name: categoryId
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Products by category.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /orders:
    post:
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "409":
          $ref: "#/components/responses/ConflictError"
  /orders/{id}:
    get:
      operationId: getOrder
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Order found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFoundError"
components:
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: integer
        minimum: 1
  responses:
    ValidationError:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFoundError:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ConflictError:
      description: Conflict.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid request payload
            details:
              type: array
              items:
                type: object
    Address:
      type: object
      required: [id, label, street, number, district, city, state, postalCode, isDefault]
      properties:
        id:
          type: integer
        label:
          type: string
        street:
          type: string
        number:
          type: string
        complement:
          type: string
          nullable: true
        district:
          type: string
        city:
          type: string
        state:
          type: string
        postalCode:
          type: string
        isDefault:
          type: boolean
    Customer:
      type: object
      required: [id, fullName, email, documentNumber, status, address, createdAt, updatedAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
          format: email
        documentNumber:
          type: string
        phone:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive]
        address:
          $ref: "#/components/schemas/Address"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CustomerPage:
      type: object
      required: [page, pageSize, total, items]
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
    CreateCustomerRequest:
      type: object
      required: [fullName, email, documentNumber, address]
      properties:
        fullName:
          type: string
        email:
          type: string
          format: email
        documentNumber:
          type: string
        phone:
          type: string
        address:
          $ref: "#/components/schemas/AddressInput"
    UpdateCustomerRequest:
      type: object
      required: [fullName, status, address]
      properties:
        fullName:
          type: string
        phone:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive]
        address:
          $ref: "#/components/schemas/AddressInput"
    AddressInput:
      type: object
      required: [label, street, number, district, city, state, postalCode, isDefault]
      properties:
        label:
          type: string
        street:
          type: string
        number:
          type: string
        complement:
          type: string
          nullable: true
        district:
          type: string
        city:
          type: string
        state:
          type: string
        postalCode:
          type: string
        isDefault:
          type: boolean
    Product:
      type: object
      required: [id, categoryId, sku, name, unitPrice, stockQuantity, active]
      properties:
        id:
          type: integer
        categoryId:
          type: integer
        sku:
          type: string
        name:
          type: string
        unitPrice:
          type: string
        stockQuantity:
          type: integer
        active:
          type: boolean
    ProductList:
      type: object
      required: [categoryId, items]
      properties:
        categoryId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
    CreateOrderRequest:
      type: object
      required: [customerId, addressId, items, payment]
      properties:
        customerId:
          type: integer
        addressId:
          type: integer
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [productId, quantity]
            properties:
              productId:
                type: integer
              quantity:
                type: integer
                minimum: 1
        payment:
          type: object
          required: [method]
          properties:
            method:
              type: string
              enum: [credit_card, debit_card, pix, boleto]
    Order:
      type: object
      required: [id, status, totalAmount, customer, address, items, payment, createdAt, updatedAt]
      properties:
        id:
          type: integer
        status:
          type: string
        totalAmount:
          type: string
        customer:
          $ref: "#/components/schemas/Customer"
        address:
          $ref: "#/components/schemas/Address"
        items:
          type: array
          items:
            type: object
        payment:
          type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time